Add feature to return taxation percent for employee and state. Add feature to calculate the incomes by year. Add new api tests.

+++ src/database/db_backup.rs
@@ -1,7 +1,7 @@
 use chrono::{Datelike, Local};
-pub fn get_current_year() -> i32 {
+pub fn get_current_year() -> u32 {
     let dt = Local::now();
-    dt.year()
+    dt.year() as u32
 }
 
 pub fn get_initial_insert_statements() -> Vec<String> {
